[#12208] Add UnsignedByte Util

+++ collector/src/main/java/com/navercorp/pinpoint/collector/mapper/grpc/GrpcJvmInfoBoMapper.java
@@ -36,7 +36,7 @@ public class GrpcJvmInfoBoMapper {
     }
 
     public JvmInfoBo map(final PJvmInfo jvmInfo) {
-        final short version = (short) jvmInfo.getVersion();
+        final int version = jvmInfo.getVersion();
         final String jvmVersion = jvmInfo.getVmVersion();
         final String gcTypeName = this.jvmGcTypeMapper.map(jvmInfo.getGcType()).name();
         return new JvmInfoBo(version, jvmVersion, gcTypeName);
